fix: use new branching strategy to reduce checkpoints size (#2048)
* rebased to main

* Added files from patch

* restored ClineProvider call to Migration deletes, added try try catch

* Restored changes to Cline.ts

* Updated tests

* Updated tests, prettier

* Optimize commit operation

* Fixed branch deletion, cleanup

* Cleanup

* minor cleanuop

* changeset

* Fix checkpoint tracker initial commit not tracking any files

* Ensure we are getting hash from branch and not main

* Revert first commit logic

* Change to get the correct hash for 'See new changes' on new and resumed tasks

* Make checkpoints commits async

* Update .changeset/cold-beers-laugh.md


---------
