+++ .changeset/cold-beers-laugh.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+Removed legacy support from checkpoints, optimized checkpoint operations
